refactor: codebase to typescript
