+++ src/core/CollectionIndex.ts
@@ -2,21 +2,23 @@ const parseAndUpdate = require('./operators/UpdateOperators')
 const parseAndFind = require('./operators/QueryOperators')
 
 class CollectionIndex {
+    _index: {};
+    loaded: boolean;
     constructor() {
         this._index = {}
         this.loaded = false;
     }
-    async find(query, projection, options = {}, callback) {
-        let res = parseAndFind(query, options, this._index, false)
+    async find(query?, projection?, options: any = {}, callback?) {
+        const res = parseAndFind(query, options, this._index, false)
         return res
     }
 
-    async findOne(query, projection, options = {}, callback) {
-        let res = parseAndFind(query, options, this._index, true)
+    async findOne(query, projection, options: any = {}, callback) {
+        const res = parseAndFind(query, options, this._index, true)
         return res
     }
 
-    async findById(_id, projection, options = {}, callback) {
+    async findById(_id, projection, options: any = {}, callback) {
         return this._index[_id]
     }
 
@@ -24,14 +26,14 @@ class CollectionIndex {
         if (!key) {
             throw "Key must not be undefined"
         }
-        var index;
+        let index;
         if (query) {
             index = await this.find(query);
         } else {
             index = this._index;
         }
-        var out = {};
-        for (var entry of Object.values(index)) {
+        const out = {};
+        for (const entry of Object.values(index)) {
             //TODO: Allow for sub object keys. I.e key.subkey
             if (entry[key]) {
                 out[entry[key]] = null;
@@ -40,7 +42,7 @@ class CollectionIndex {
         return Object.keys(out);
     }
     handleEntry(item) {
-        var { payload } = item;
+        const { payload } = item;
         switch (payload.op) {
             case "INSERT": {
                 this.handleInsert(payload);
@@ -69,21 +71,21 @@ class CollectionIndex {
         }
     }
     async handleInsert(payload) {
-        var { value } = payload;
-        for (var doc of value) {
-            var _id = doc._id;
+        const { value } = payload;
+        for (const doc of value) {
+            const _id = doc._id;
             this._index[_id] = doc;
         }
     }
     handleUpdate(payload) {
-        var { value, modification, options } = payload;
-        for (var _id of value) {
+        const { value, modification, options } = payload;
+        for (const _id of value) {
             parseAndUpdate(this._index[_id], modification, options)
         }
     }
     handleDelete(payload) {
-        var { value } = payload;
-        for(var _id of value) {
+        const { value } = payload;
+        for(const _id of value) {
             delete this._index[_id];
         }
     }
