+++ tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "outDir": "./dist/",
+    "lib": ["es2018", "dom", "dom.iterable"],
+    "module": "commonjs",
+    "target": "es6"
+  }
+}
